pbl-openai-be/pyproject.toml
[tool.poetry]
name = "pbl-openai-be"
version = "0.1.0"
description = ""
authors = ["yixing wang <yixwang@ea.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
openai = "^1.28.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

